pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "HARIS"
version = "0.1.0"
description = "Black-box web application security audit framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
authors = [
    {name = "HARIS contributors"},
]
keywords = ["security", "audit", "pentest", "owasp", "vulnerability-scanner"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security",
]

dependencies = [
    "pydantic>=2.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "openai>=2.21.0",
    "anthropic>=0.83.0",
]

[project.optional-dependencies]
# Web dashboard
web = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
]
# Report generation extras
reporting = [
    "markdown>=3.5",
]
# LLM adapter (choose one or more provider packages)
llm-openai = [
    "openai>=1.0.0",
]
llm-anthropic = [
    "anthropic>=0.20.0",
]
# Development / testing
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.7",
    "types-requests",
    "types-PyYAML",
]
# Scanner Python packages (optional — scanners degrade gracefully if missing)
scanners = [
    "sslyze>=6.0.0",
    "wapiti3>=3.1.0",
]
# Everything
all = [
    "HARIS[web,reporting,llm-anthropic,scanners,dev]",
]

[project.scripts]
HARIS = "scripts.run_scan:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
